+++ api-testcases/test-push-global.c
@@ -6,7 +6,7 @@ duk_is_function(global.escape)=1
 
 void test(duk_context *ctx) {
 	duk_push_global_object(ctx);
-	printf("type=%d\n", duk_get_type(ctx, -1));
+	printf("type=%d\n", (int) duk_get_type(ctx, -1));
 	printf("ToString=%s\n", duk_to_string(ctx, -1));
 	duk_pop(ctx);
